\name{EuclideanMaps-package}
\Rdversion{1.1}
\alias{EuclideanMaps-package}
\docType{package}
\title{Euclidean Maps}
\description{
Displays Euclidean Maps
}
\details{
\tabular{ll}{
Package: \tab MapasEuclideos\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-12-09\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Elisa Frutos Bernal \email{efb@usal.es}
}
\keyword{ package }
