\name{EuclideanMaps}
\Rdversion{1.1}
\alias{EuclideanMaps}
\title{
Euclidean Maps
}
\description{
Displays Row and Column Euclideans Maps
}
\usage{
EuclideanMaps()
}
\details{
From the data matrix are calculated the coordinates for points representing row and column respectively 
to the rows and columns of the same. The row Euclidean map is a graph in the cartesian plane row points 
and column Euclidean map is a graph in the Cartesian plane column points.                                                                        
}
\value{
Returns Row Euclidean Map and Column Euclidean Map
}
\references{
Bradu D, Gabriel KR.(1978). The biplot as a diagnostic tool for models of two-ways tables. Technometrics, 20, 47-68.
Bradu D (1983). Model Diagnosis in two-way tables by means of row and column Euclidean Maps. Tecnical report TWISK 307. National Research Institute  for Mathematical Sciences. Pretoria.
Gabriel KR (1971). The biplot graphic display of matrices with applications to principal component analysis. Biometrica, 58, 453-467.
Gabriel KR (1978). Least Squares Approximation of Matrices by Additive Models. Journal of Royal Statistic Society, 40:186-196.
Gabriel KR (1981). Biplot Display of Multivariate Matrices for Inspection of Data and Diagnosis. In: V. Barnett (Ed.) Interpreting Multivariate Data, 58: 453-467.
}
\author{
Elisa Frutos Bernal
}
\examples{
EuclideanMaps()
}
\keyword{ models }
